Use createAgGridColDef "progress" property (just a callback, no need to import ProgressCell)
