+++ .changeset/large-poets-allow.md
@@ -1,0 +1,5 @@
+---
+'@platforma-open/milaboratories.mixcr-clonotyping.ui': patch
+---
+
+Use createAgGridColDef "progress" property (just a callback, no need to import ProgressCell)
