+++ pnpm-workspace.yaml
@@ -8,7 +8,7 @@ packages:
 catalog:
   '@platforma-sdk/workflow-tengo': ^2.9.9
   '@platforma-sdk/model': ^1.21.20
-  '@platforma-sdk/ui-vue': ^1.22.9
+  '@platforma-sdk/ui-vue': ^1.22.39
   '@platforma-sdk/tengo-builder': ^1.17.5
   '@platforma-sdk/block-tools': ^2.5.3
   '@platforma-sdk/eslint-config': ^1.0.1
